pyproject.toml
[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[project]
name = "attpc_conduit"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
	"cmap>=0.1.3",
    "dearpygui>=1.10",
    "rerun-sdk>=0.16.0",
    "h5py>=3.10",
    "attpc_spyral>=0.6.1"
]

[project.scripts]
run-conduit = "attpc_conduit.main:main"
rerun-loader-merged-file = "attpc_conduit.rerun_loader_merged_file:main"

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "attpc_conduit._attpc_conduit"
python-source = "python"
profile = "release"
